# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create/payment": {
            "post": {
                "tags": [
                    "Payment Section"
                ],
                "summary": "Request prepares payments",
                "description": "",
                "operationId": "f08d56b1fee3cdeb42404409cf6339f2",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/payment": {
            "get": {
                "tags": [
                    "Payment Section"
                ],
                "summary": "Request which returns payments",
                "description": "",
                "operationId": "f17cae43e272cae94da01aa9ea311b64",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/accept/purchase": {
            "post": {
                "tags": [
                    "Payment Section"
                ],
                "summary": "Request confirmation",
                "description": "",
                "operationId": "60d4dafbd8529e575211b1ee6b035682",
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "query",
                        "description": "Please write payment ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Please write status",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}